Bare minimum needed to link the new SBBM backend.
As far as I can tell, this is the minimum number of build system changes
and source files necessary to call yourself a backend.  The backend
doesn't do anything except report that it exists.  Any attempt to invoke a
tool using the SBBM backend will result in crashes.

Three sub-projects are introduced: CodeGen (the main projects),
MCTargetDesc, and TargetInfo.  At this stage, these three are not strictly
necessary, but this pattern is followed by other backends.  The separation
will make more sense as code is added.