Adds support for sub instructions.
Also adds transformations from ADDri and SUBri with negative immediate
to the other one with positive immediate.

NOTE: The the ADDri neg -> SUBri pos is necessary at this
stage (desirable regardless) because LLVM applies an unconditional
optimization of sub with immediate to add with opposite sign immediate.

NOTE: The test of non-commutability in sub.ll cannot be executed because
copyPhysReg is required to swap the values.