Workarounds to get more test code compiling.
Changes the number of general registers to an astonishing 32, so that we
can avoid saving values onto the stack (which does not exist yet).  The
calling convention is also change to pass the first 16(!) values to a
function in registers and return up to 16(!!) values in registers.  This
behavior will probably be desirable even after the memory controller is
available, since any implementation of memory in Minecraft is likely to
be very slow.

Adds 'support' for addc, adde, subc, sube.  They are just redirected to
ordinary add/sub for now.

Adds 'support' for udiv, urem.  Again, just redirect to sdiv, srem.

Tells LLVM to expand quite a few more instructions, so we can avoid
implementing them for now.