Adds def 'macros' to aid assembly generation.
Because SBBM's assembly is unusual (it requires quite a lot of
static string-building to pass raw commands to Minecraft), this adds
def-combiner and def-expander passes to the compilation process.

def-combiner is basically a (very) poor man's macro system.  If you
write:

    def foo, hello
    raw say %foo

The result is:

    raw say hello

def-remover just goes around removing defs after they have been
expanded.  The removal can't happen as expansion is ocurring because
there is no easy way to know when a function will never be visited
again.  A separate pass eliminates this concern.

There is nothing terribly nice about this, but it had to be done.