Print predicates first thing on the line.
This is required if raw commands are ever going to be prediable (they
are), since it would be difficult to both treat raw command as a black
box *and* extract the predicate from the end.  This way, the predicate
is handled first, and then the rest of the line can be freeform.